+++ web-local/src/routes/(viz)/explore/[name]/+page.svelte
@@ -9,7 +9,7 @@
   import DashboardThemeProvider from "@rilldata/web-common/features/dashboards/DashboardThemeProvider.svelte";
   import { resetSelectedMockUserAfterNavigate } from "@rilldata/web-common/features/dashboards/granular-access-policies/resetSelectedMockUserAfterNavigate";
   import { selectedMockUserStore } from "@rilldata/web-common/features/dashboards/granular-access-policies/stores";
-  import DashboardURLStateSync from "@rilldata/web-common/features/dashboards/url-state/DashboardURLStateSync.svelte";
+  import DashboardStateManager from "@rilldata/web-common/features/dashboards/state-managers/loaders/DashboardStateManager.svelte";
   import StateManagersProvider from "@rilldata/web-common/features/dashboards/state-managers/StateManagersProvider.svelte";
   import { useProjectParser } from "@rilldata/web-common/features/entity-management/resource-selectors";
   import { useExploreValidSpec } from "@rilldata/web-common/features/explores/selectors";
@@ -19,28 +19,7 @@
   import { queryClient } from "@rilldata/web-common/lib/svelte-query/globalQueryClient";
 
   export let data: PageData;
-  $: ({
-    metricsView,
-    explore,
-    defaultExplorePreset,
-    exploreStateFromYAMLConfig,
-    partialExploreStateFromUrl,
-    exploreStateFromSessionStorage,
-    errors,
-    exploreName,
-  } = data);
-  $: if (errors?.length) {
-    const _errs = errors;
-    setTimeout(() => {
-      eventBus.emit("notification", {
-        type: "error",
-        message: _errs[0].message,
-        options: {
-          persisted: true,
-        },
-      });
-    }, 100);
-  }
+  $: ({ metricsView, explore, exploreName } = data);
 
   resetSelectedMockUserAfterNavigate(queryClient);
 
@@ -111,18 +90,11 @@
 {:else}
   {#key exploreName}
     <StateManagersProvider {metricsViewName} {exploreName}>
-      <DashboardURLStateSync
-        {metricsViewName}
-        {exploreName}
-        {defaultExplorePreset}
-        {exploreStateFromYAMLConfig}
-        {partialExploreStateFromUrl}
-        {exploreStateFromSessionStorage}
-      >
+      <DashboardStateManager {exploreName}>
         <DashboardThemeProvider>
           <Dashboard {metricsViewName} {exploreName} />
         </DashboardThemeProvider>
-      </DashboardURLStateSync>
+      </DashboardStateManager>
     </StateManagersProvider>
   {/key}
 {/if}
